refactor: extract shared postSlackMessage module for Lambda workers (#140)
Create lambda/worker/shared/slack.ts with a common postSlackMessage
function using raw fetch, replacing three duplicate implementations
in question.ts, question-followup.ts, and issue.ts.
